REFACTOR tests classes into `Feature` namespace
