+++ tests/Feature/Pdf/PdfExportFromHtmlFileTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Sfneal\ViewExport\Tests\Pdf;
+namespace Sfneal\ViewExport\Tests\Feature\Pdf;
 
 use Sfneal\ViewExport\Pdf\PdfExportService;
 
-class PdfExportFromUrlTest extends PdfTestCase
+class PdfExportFromHtmlFileTest extends PdfTestCase
 {
     /**
      * Setup the test environment.
@@ -15,6 +15,6 @@ class PdfExportFromUrlTest extends PdfTestCase
     {
         parent::setUp();
 
-        $this->renderer = PdfExportService::fromHtmlFile('https://example.com/');
+        $this->renderer = PdfExportService::fromHtmlFile(base_path('tests/resources/html/test.html'));
     }
 }
