--- tests/Pdf/PdfExportFromViewTest.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Sfneal\ViewExport\Tests\Pdf;
-
-use Sfneal\ViewExport\Pdf\PdfExportService;
-
-class PdfExportFromViewTest extends PdfTestCase
-{
-    /**
-     * Setup the test environment.
-     *
-     * @return void
-     */
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->renderer = PdfExportService::fromView(view('test'));
-    }
-}
